+++ glacier_flow_model/fracd8/limited.py
@@ -1,0 +1,63 @@
+from typing import Tuple
+
+import numpy as np
+from numba import njit
+
+
+@njit
+def fracd8_lim(
+    ele: np.ndarray, u: np.ndarray, h: np.ndarray, res: float
+) -> Tuple[np.ndarray, np.ndarray]:
+
+    # Setup arrays
+    asp = np.zeros_like(ele, dtype=np.uint8)
+    asp[:] = np.nan
+    h_flow = np.zeros_like(ele, dtype=np.float32)
+
+    # Iterate through grid
+    rows, cols = ele.shape
+    for y in range(1, rows - 1):
+        for x in range(1, cols - 1):
+            pos = [
+                (y, x),  # 0
+                (y + 1, x + 1),  # 1
+                (y, x + 1),  # 2
+                (y - 1, x + 1),  # 3
+                (y - 1, x),  # 4
+                (y - 1, x - 1),  # 5
+                (y, x - 1),  # 6
+                (y + 1, x - 1),  # 7
+                (y + 1, x),  # 8
+            ]
+            pos_0 = pos[0]
+
+            # Classify aspect: Get steepest dz
+            ele_0 = ele[pos_0]
+            asp_0 = 0
+            dz = 0
+            for i in range(1, 9):
+                dz_tmp = ele_0 - ele[pos[i]]
+                if dz_tmp > dz:
+                    asp_0 = i
+                    dz = dz_tmp
+            asp[pos_0] = asp_0
+
+            # Flow
+            h_0 = h[pos_0]
+            if h_0 < 0.00001:
+                # Nothing to flow!
+                continue
+
+            if asp_0 == 0:
+                # Nowhere to flow!
+                h_flow[pos_0] += h_0
+                continue
+
+            # Velocity defines fraction that stays in pixel
+            u_0 = u[pos_0]
+            fraction = u_0 / res
+            pos_i = pos[asp_0]
+            h_flow[pos_0] += h_0 * (1 - fraction)
+            h_flow[pos_i] += h_0 * fraction
+
+    return h_flow, asp